Fix test

+++ test/yosupo/set_power_series/subset_convolution.test.cpp
@@ -14,7 +14,7 @@ int main(void) {
     for (auto &e : b) std::cin >> e;
     auto ans = subset_convolution(a, b);
     for (int i = 0; i < (int)ans.size(); ++i)
-        std::cout << ans[i] << (i == ans.size() - 1 ? '\n' : ' ');
+        std::cout << ans[i] << (i == (int)ans.size() - 1 ? '\n' : ' ');
 
     return 0;
 }
